Targeting Page ID
This CSS code lets you change something on a specific page using the page ID and class/id of what you want to alter on that page. I found that the page-id-(Page ID) needs to be a class selector, not an id.

I used this code on a WordPress website. To find the page id I did 'view source'. The page id is located in the 'body tag' as a class.

<body class="home page page-id-28"> for example.

This is perfect when working with classes or ids that you can't access directly.